Отрефакторил код что бы было читаемо и в одной стилистике
