+++ src/main/java/ru/chariot/dictionary/model/Dictionary.java
@@ -1,16 +1,17 @@
 package ru.chariot.dictionary.model;
 
-
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.persistence.*;
-
-import lombok.Data;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import jakarta.persistence.Id;
+import jakarta.persistence.Column;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 
-
 @Entity
-@Data
+@lombok.Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Table(name = "dictionary")
@@ -21,10 +22,13 @@ public class Dictionary {
     @Column(name = "id" , nullable = false, unique = true)
     @Schema(name = "id", hidden = true)
     private Long id;
+
     @Column(name = "code", nullable = false)
     @Schema(name = "code", example = "test", required = true)
     private String code;
+
     @Column(name = "description", nullable = true)
     @Schema(name = "description", example = "test", required = true)
     private String description;
+
 }
