+++ src/main/java/ru/chariot/dictionary/reposetory/DataRepository.java
@@ -2,8 +2,8 @@ package ru.chariot.dictionary.reposetory;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.chariot.dictionary.model.Dictionary;
+import ru.chariot.dictionary.model.Data;
 
 @Repository
-public interface DictionaryReposetory extends JpaRepository<Dictionary, Long> {
+public interface DataRepository extends JpaRepository<Data, Long> {
 }
