+++ src/main/java/ru/chariot/dictionary/reposetory/DictionaryRepository.java
@@ -2,8 +2,8 @@ package ru.chariot.dictionary.reposetory;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.chariot.dictionary.model.Data;
+import ru.chariot.dictionary.model.Dictionary;
 
 @Repository
-public interface DataReposetory extends JpaRepository<Data, Long> {
+public interface DictionaryRepository extends JpaRepository<Dictionary, Long> {
 }
